+++ website/demo.js
@@ -1,6 +1,5 @@
-/* global TrackyMouse */
-
 import { InputSimulator } from "./core/src/input-simulator.js";
+import { TrackyMouse } from "./core/src/tracky-mouse.js";
 
 TrackyMouse.dependenciesRoot = "./core";